Decouple business logic and IO
